pyproject.toml
[project]
name = "property-insights-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "ipywidgets>=8.1.7",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-openai>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
]
